backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
requests
markdown
python-dotenv
